==> analyses/02-evaluate_comprehensiveness.Rmd <==
---
title: "02-evaluate_comprehensiveness"
author: "Devi Veytia"
date: "2023-01-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

```{r load libraries}
library(dplyr)
library(stringi)
library(stringr)
```

```{r set up inputs}

## file locations for search string
search_string_dir <- here::here("data","derived-data","search-string")
search_string_fp <- file.path(search_string_dir,"search-string-blocks.xlsx")


```




Using the test list, the search strings and the search results, I want to evaluate the comprehensiveness of the search string.


# Format the search string

Using the search string components (formula, keywords and substrings) stored in the search string file, build the full search string and format for Scopus

## For Scopus


```{r search string format for scopus, echo=FALSE}

## Set up

# get a list of all the sheets to iterate over
sheets <- readxl::excel_sheets(path = search_string_fp)

# set up a dataframe to store the string results
string_blocks_scopus <- data.frame(Block = sheets)
string_blocks_scopus$String <- NA


## Loop through the sheets and read in and format the string

for(s in 1:length(sheets)){
  
  ## read in the search string for each block located within each sheet
  # read in file
  temp <- readxl::read_excel(path = search_string_fp, sheet = sheets[s], trim_ws = TRUE, skip = 2)
  temp <- temp %>% select(-c(Date, `Search #`, Block, `Test list results`, `Total search results`, Comments))
  
  # fill in empty columns with boolean operator
  for(c in 1:ncol(temp)){
    if(sum(is.na(temp[,c])) == nrow(temp)){
      temp[,c] <- str_remove_all(colnames(temp)[c], "[^[:alpha:]]") # have to clean out unique suffix of column name e.g. "...5"
    }
  }
  
  # to handle "" fill with the value from the previous row
  for(i in 1:dim(temp)[1]){
    for(j in 1:dim(temp)[2]){
      if(is.na(temp[i,j])){next}
      if(temp[i,j] == "\"\""){
        temp[i,j] <- temp[i-1,j]
      }else{
        next
      }
    }
  }
  
  # the last row is where the final string is stored
  string <- temp[nrow(temp),]
  
  # clean out an NA and preceeding boolean operator if it is present
  if(sum(is.na(string))>0){
    naInd <- which(is.na(string))
    rmOper <- naInd+1
    string[-c(naInd, rmOper)]
  }
  
  
  ## Format the string for Scopus
  
  # replace NEAR/x operators
  string <- unlist(gsub("NEAR","W",string))
  
  # replace "NOT" with "AND NOT"
  string <- unlist(gsub("NOT", "AND NOT",string))
  
  # Replace with the scopus field codes
  string <- unlist(gsub("TS = ", "TITLE-ABS-KEY",string))
  string <- unlist(gsub("LA=", "LANGUAGE",string))
  
  # combine seperate elements together into one string
  string <- paste0(string, collapse =" ")
  
  # store
  string_blocks_scopus$String[s] <- string
  
}


```


```{r write search equation for scopus to txt file}

# # add in field codes and write to text file
# 
# sink(file.path(search_string_dir,"scopus-equation-blocks.txt"))
# for(s in 1:length(sheets)){
#   cat(string_blocks_scopus$Block[s],"\n",string_blocks_scopus$String[s],"\n","\n")
# }
# sink()
```

For Scopus it might make the most sense to download records by source title, becaues there are only 4 records that exceed 2,000 references




# Evaluate comprehensiveness of the search in WOS 


Using the test list, evaluate the comprehensiveness of the search in WOS. 

The test list items retrieved from each sub-string (adding the additional argument in the search "AND (DO = <test list article dois>)" and then exporting the results), are then compiled and assessed for completeness -- is the whole test list retreived across all the sub-strings?



```{r combine test list items returned from each block}
returnTestListDir <- here::here("data","raw-data","WOS_downloads","test-list-returns-by-block")
returnTestListFiles <- dir(returnTestListDir)

# read in and de-duplicate test list returns
temp <- rbibtools::read_bib(here::here(returnTestListDir))
temp$unique_id <- revtools::find_duplicates(temp)
temp <- temp[!duplicated(temp$unique_id),]
nrow(temp)

# compare to test list
# read in test list
test_list <- readxl::read_excel(here::here("data", "derived-data", "screen_test_list", "test-list_2023-01-19.xlsx"))

# identify if any test list items are not included
ind <- which(!(test_list$doi %in% temp$doi))
if(length(ind)>0){
  View(test_list[ind,])
}else{
  print("all test list articles retreived")
}



```
